Agno/agno_finance_agent/agno_finance_agent_ai/pyproject.toml
[project]
name = "agno-finance-agent-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.5",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "packaging>=25.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
